generator: move NE populated places code to separate file
